+++ src/cljs/rems/actions/vote.cljs
@@ -48,7 +48,8 @@
       {:id id
        ;; XXX: consider making choices dynamic
        :items [{:value "accept" :label (text :t.applications.voting.votes/accept)}
-               {:value "reject" :label (text :t.applications.voting.votes/reject)}]
+               {:value "reject" :label (text :t.applications.voting.votes/reject)}
+               {:value "empty" :label (text :t.applications.voting.votes/empty)}]
        :item-label (comp rc/as-element :label)
        :item-selected? #(= (or vote previous-vote) (:value %))
        :on-change #(on-vote (:value %))}]]))
@@ -92,37 +93,62 @@
                                                      :attachments attachments
                                                      :on-finished on-finished}])}]))
 
+;; TODO too difficult, consider changing data model
+;; TODO consider moving to a util
+(defn- get-reviewers [application]
+  (for [event (:application/events application)
+        :let [type (:event/type event)]
+        :when (= type :application.event/review-requested)
+        reviewer (:application/reviewers event)]
+    reviewer))
+
+(defn- get-potential-voters [application]
+  (case (-> application :application/workflow :workflow/voting :type)
+    :handlers-vote
+    (get-in application [:application/workflow :workflow.dynamic/handlers])
+
+    :reviewers-vote
+    (get-reviewers application)
+
+    nil))
+
 (defn votes-summary
   [application]
   (let [votes (get-in application [:application/votes])
         summary (frequencies (vals votes))
-        voters (get-in application [:application/workflow :workflow.dynamic/handlers])
+        voters (get-potential-voters application)
         voter-by-userid (build-index {:keys [:userid]} voters)
         voters-by-vote (build-index {:keys [val] :value-fn (comp voter-by-userid key) :collect-fn conj} votes)]
-    [:div.my-3
-     [:h3 (text :t.applications/votes)]
+    (when (seq voters)
+      [:div.my-3
+       [:h3 (text :t.applications/votes)]
 
-     [:div.container-fluid
-      (for [[vote n] (sort-by val summary)
-            :let [n-pct (* 100 (/ n (count voters)))
-                  vote-voters (get voters-by-vote vote)]]
-        ^{:key vote}
-        [:div.form-group.row
-         [:label.col-sm-3.col-form-label (text (keyword (str "t" ".applications.voting.votes") vote))]
-         [:div.col-sm-9.form-control (goog.string/format "%.2f%% (%s)"
-                                                         n-pct
-                                                         (->> vote-voters
-                                                              (mapv application-util/get-member-name)
-                                                              (str/join ", ")))]])
+       [:div.container-fluid
+        (into [:<>] (for [[vote n] (sort-by val summary)
+                          :let [n-pct (* 100 (/ n (count voters)))
+                                vote-voters (->> vote
+                                                 (get voters-by-vote)
+                                                 (remove nil?))]
+                          :when (seq vote-voters)
+                          :let [voters-text (->> vote-voters
+                                                 (mapv application-util/get-member-name)
+                                                 (str/join ", "))]]
+                      [:div.form-group.row
+                       [:label.col-sm-3.col-form-label (text (keyword (str "t" ".applications.voting.votes") vote))]
+                       [:div.col-sm-9.form-control (goog.string/format "%.2f%% (%s)"
+                                                                       n-pct
+                                                                       voters-text)]]))
 
-      (let [n (- (count voters) (count votes))
-            n-pct (* 100 (/ n (count voters)))
-            missing-voters (->> voters
-                                (remove (comp (set (keys votes)) :userid))
-                                (mapv application-util/get-member-name)
-                                (str/join ", "))]
-        [:div.form-group.row
-         [:label.col-sm-3.col-form-label (text :t.applications.voting.votes/empty)]
-         [:div.col-sm-9.form-control (goog.string/format "%.2f%% (%s)"
-                                                         n-pct
-                                                         missing-voters)]])]]))
+        (let [n (- (count voters) (count votes))
+              n-pct (* 100 (/ n (count voters)))
+              missing-voters (remove (comp (set (keys votes)) :userid) voters)
+              voters-text (->> missing-voters
+                               (mapv application-util/get-member-name)
+                               (str/join ", "))]
+          (when (seq missing-voters)
+            ^{:key "not-voted"}
+            [:div.form-group.row
+             [:label.col-sm-3.col-form-label (text :t.applications.voting.votes/not-voted)]
+             [:div.col-sm-9.form-control (goog.string/format "%.2f%% (%s)"
+                                                             n-pct
+                                                             voters-text)]]))]])))